+++ adminer/select.inc.php
@@ -410,10 +410,11 @@ if (!$columns && support("table")) {
 				$unique_idf = "";
 				foreach ($unique_array as $key => $val) {
 					$field = (array) $fields[$key];
-					if ((JUSH == "sql" || JUSH == "pgsql") && preg_match('~char|text|enum|set~', $field["type"]) && strlen($val) > 64) {
+					$is_binary = is_blob($field); // binary and varbinary are converted to hexadecimal so they are not shortened
+					if ((JUSH == "sql" || JUSH == "pgsql") && ($is_binary || preg_match('~char|text|enum|set~', $field["type"])) && strlen($val) > 64) {
 						$key = (strpos($key, '(') ? $key : idf_escape($key)); //! columns looking like functions
-						$key = "MD5(" . (JUSH != 'sql' || preg_match("~^utf8~", $field["collation"]) ? $key : "CONVERT($key USING " . charset(connection()) . ")") . ")";
-						$val = md5($val);
+						$key = "MD5(" . ($is_binary || JUSH != 'sql' || preg_match("~^utf8~", $field["collation"]) ? $key : "CONVERT($key USING " . charset(connection()) . ")") . ")";
+						$val = md5($is_binary ? (string) driver()->value($val, $field) : $val); // value() decodes bytea in PostgreSQL
 					}
 					$unique_idf .= "&" . ($val !== null ? urlencode("where[" . bracket_escape($key) . "]") . "=" . urlencode($val === false ? "f" : $val) : "null%5B%5D=" . urlencode($key));
 				}